jdbc almost done, need test

+++ jdbc/src/repositories/BaseRepository.java
@@ -8,8 +8,8 @@ import entities.IEntity;
 
 public interface BaseRepository {
     List<IEntity> getAll(Connection connection) throws SQLException;
-    IEntity getById(Connection connection, int id) throws SQLException;
+    IEntity getById(Connection connection, Long id) throws SQLException;
     void insert(Connection connection, IEntity entity) throws SQLException;
-    void delete(Connection connection, int id) throws SQLException;
+    void delete(Connection connection, Long id) throws SQLException;
     void update(Connection connection, IEntity entity) throws SQLException;
 }
